SEQNG-58 Ignore pending tests
